requirements.txt
# backend/requirements.txt - UPDATED
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
opencv-python==4.8.1.78
numpy==1.24.3
google-generativeai==0.3.2
python-dotenv==1.0.0
vosk==0.3.45
psutil==5.9.6
aiofiles==23.2.1
moviepy==1.0.3
pydub==0.25.1
SpeechRecognition==3.10.0
mediapipe==0.10.11
google-cloud-speech==2.21.0
SpeechRecognition==3.10.0
ffmpeg-python==0.2.0
